+++ finops_watchdog/__init__.py
@@ -1,0 +1,3 @@
+"""FinOps Watchdog."""
+
+__version__ = "0.2.0"